build: Updated file structure This is so that file structure would not be clumsy
