--- src/app/layout.js
@@ -1,17 +1,0 @@
-import "./globals.css";
-import { Inter } from "next/font/google";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata = {
-	title: "Create Next App",
-	description: "Generated by create next app",
-};
-
-export default function RootLayout({ children }) {
-	return (
-		<html lang="en">
-			<body className={inter.className}>{children}</body>
-		</html>
-	);
-}
