--- src/app/page.js
@@ -1,12 +1,0 @@
-import Header from "./Components/Header";
-import FindAMentor from "./findamentor/page";
-
-const Page = () => {
-	return (
-		<>
-			<FindAMentor />
-		</>
-	);
-};
-
-export default Page;
