+++ src/pages/globals.css
@@ -1,0 +1,6 @@
+/* src/styles/globals.css */
+@import 'tailwindcss/base';
+@import 'tailwindcss/components';
+@import 'tailwindcss/utilities';
+
+/* Your custom global styles can go here */
