+++ src/pages/index.js
@@ -1,0 +1,13 @@
+import Header from "../Components/Header";
+import FindAMentor from "./findamentor";
+import { Inter } from "next/font/google";
+
+const Page = () => {
+	return (
+		<>
+			<FindAMentor />
+		</>
+	);
+};
+
+export default Page;
