+++ src/pages/partnership.js
@@ -1,16 +1,15 @@
-import Handshake from "../Components/Handshake";
-import Footer from "../Components/Footer";
-import PartnershipLeft from "../Components/PartnershipLeft";
-import PartnershipForm from "../Components/PartnershipForm";
-import Header from "../Components/Header";
+import Handshake from "@/Components/Handshake";
+import Footer from "@/Components/Footer";
+import PartnershipLeft from "@/Components/PartnershipLeft";
+import PartnershipForm from "@/Components/PartnershipForm";
+import Header from "@/Components/Header";
 import Image from "next/image";
 
 function PartnershipPage() {
 	return (
 		<div>
 			<Header />
-			<Handshake />
-
+			<Handshake /> {/* This is the correct usage of the Handshake component */}
 			<div className="flex justify-between pt-14 pl-24 pb-[400px] relative">
 				<Image
 					src="/images/right-tri.png"
@@ -22,7 +21,6 @@ function PartnershipPage() {
 				<PartnershipLeft />
 				<PartnershipForm />
 			</div>
-
 			<Footer />
 		</div>
 	);
